Serialization refactoring.
